Refactor model-edit forms, create common superclass.
