+++ opengever/meeting/browser/members.py
@@ -2,13 +2,12 @@ from five import grok
 from opengever.meeting import _
 from opengever.meeting.committeecontainer import ICommitteeContainer
 from opengever.meeting.form import ModelAddForm
+from opengever.meeting.form import ModelEditForm
 from opengever.meeting.model import Member
 from plone.directives import form
 from Products.Five.browser import BrowserView
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
-from z3c.form import button
 from z3c.form import field
-from z3c.form.form import EditForm
 from zExceptions import NotFound
 from zope import schema
 from zope.interface import implements
@@ -41,52 +40,12 @@ class AddMember(ModelAddForm):
     model_class = Member
 
 
-class EditMember(EditForm):
+class EditMember(ModelEditForm):
 
-    ignoreContext = True
     fields = field.Fields(IMemberModel)
-
-    is_model_view = True
-    is_model_edit_view = True
-
-    def __init__(self, context, request, model):
-        super(EditMember, self).__init__(context, request)
-        self.model = model
-        self._has_finished_edit = False
-
-    def inject_initial_data(self):
-        if self.request.method != 'GET':
-            return
-
-        prefix = 'form.widgets.'
-        values = self.model.get_edit_values(self.fields.keys())
-
-        for fieldname, value in values.items():
-            self.request[prefix + fieldname] = value
-
-    def updateWidgets(self):
-        self.inject_initial_data()
-        super(EditMember, self).updateWidgets()
-
-    def applyChanges(self, data):
-        self.model.update_model(data)
-        # pretend to always change the underlying data
-        self._has_finished_edit = True
-        return True
-
-    # this renames the button but otherwise preserves super's behavior
-    @button.buttonAndHandler(_('Save'), name='save')
-    def handleApply(self, action):
-        # self as first argument is required by the decorator
-        super(EditMember, self).handleApply(self, action)
 
     def nextURL(self):
         return MemberView.url_for(self.context, self.model)
-
-    def render(self):
-        if self._has_finished_edit:
-            return self.request.response.redirect(self.nextURL())
-        return super(EditMember, self).render()
 
 
 class MemberList(grok.View):
